telegram/requirements.txt
aiogram==3.7.0
pika==1.3.2
ujson==5.10.0
requests==2.32.3
psycopg2-binary==2.9.9